Branch into rendering different panels based on the current route
